re PR fortran/32770 ([Meta-bug] -fdefault-integer-8 issues)
2007-08-03  Thomas Koenig  <tkoenig@gcc.gnu.org>

	PR fortran/32770
	* gfortran.dg/unf_read_corrupted_1.f90:  Integer variable in
	iostat must be default kind.

From-SVN: r127178

+++ gcc/testsuite/gfortran.dg/unf_read_corrupted_1.f90
@@ -4,7 +4,7 @@
 program main
   implicit none
   integer(kind=4) :: i1, i2
-  integer(kind=4) :: ios
+  integer :: ios
   character(len=50) :: msg
 
   ! Write out a truncated unformatted sequential file by
